relocate Bodu.IO.Hashing-only contracts and KATs out of Bodu.Test
Three contract bases and four KAT records were consumed only by
Bodu.IO.Hashing.Test, so they move alongside their consumers per the
single-consumer rule. Two orphaned KAT records (ChecksumKat,
InvalidCheckDigitKat) had no consumers anywhere — investigated per
the per-item rule and deleted as dead code.

Moved to Bodu.IO.Hashing/test/Contracts/ (namespace
Bodu.IO.Hashing.Contracts):

Contracts:
- NonCryptographicHashAlgorithmContractTests<TAlgorithm>
- CheckDigitContractTests<TAlgorithm>
- MultiCharCheckDigitContractTests<TAlgorithm>

KATs:
- HashKat
- HashStreamingKat
- CrcCatalogKat
- CheckDigitKat

Deleted as dead code:
- ChecksumKat (no consumers anywhere)
- InvalidCheckDigitKat (no consumers anywhere)

Subclass `using` updates:
- Same-namespace subclasses (Bodu.IO.Hashing.Contracts): removed unused
  `using Bodu.Test.Contracts;` and `using Bodu.Test.Kat;`.
- Cross-namespace subclasses (Bodu.IO.Hashing.{CheckDigits,Checksums}
  .Contracts): replaced both `using` lines with a single
  `using Bodu.IO.Hashing.Contracts;`.
- Domain-local KAT records (NonCryptographicHashKnownAnswer,
  CheckDigitKnownAnswer, MultiCharCheckDigitKnownAnswer,
  MultiCharCheckDigitIsValidKnownAnswer): retain
  `using Bodu.Test.Kat;` because they implement the IKat marker.
- CrcCatalogKatTests: added `using Bodu.IO.Hashing.Contracts;` because
  it references the relocated CrcCatalogKat from a different namespace.

Domain-local KAT records inspected but not consolidated (different
abstractions than the relocated records):
- NonCryptographicHashKnownAnswer
- CheckDigitKnownAnswer
- MultiCharCheckDigitKnownAnswer
- MultiCharCheckDigitIsValidKnownAnswer

HashExtensionKat was reported by CLAUDE.md as a Bodu.IO.Hashing
record, but on inspection it is consumed only by
Bodu.Security.Cryptography.Test. It will be relocated to that project
in a follow-up commit.

Verified: dotnet build passes 0 errors;
dotnet test --settings bvt.runsettings passes 3724/3724 (181
pre-existing skips).
