+++ Bodu.IO.Hashing/test/Contracts/HashStreamingKat.cs
@@ -1,10 +1,12 @@
-﻿// ---------------------------------------------------------------------------------------------------------------
+// ---------------------------------------------------------------------------------------------------------------
 // <copyright file="HashStreamingKat.cs" company="Bodu Pty. Ltd.">
 //     Copyright (c) Bodu Pty. Ltd.. All rights reserved.
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-namespace Bodu.Test.Kat;
+using Bodu.Test.Kat;
+
+namespace Bodu.IO.Hashing.Contracts;
 
 /// <summary>
 /// Represents a streaming-parity known-answer test row: a raw byte payload fed to the algorithm via a series of
