+++ Bodu.IO.Hashing/test/IO.Hashing.CheckDigits/CheckDigitKnownAnswer.cs
@@ -25,6 +25,7 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
+
 using Bodu.Test.Kat;
 
 namespace Bodu.IO.Hashing.CheckDigits;
