+++ Bodu.IO.Hashing/test/IO.Hashing.CheckDigits/MultiCharCheckDigitKnownAnswer.cs
@@ -4,6 +4,7 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
+
 using Bodu.Test.Kat;
 
 namespace Bodu.IO.Hashing.CheckDigits;
