+++ Bodu.IO.Hashing/test/IO.Hashing.Checksums/LeiContractTests.cs
@@ -4,8 +4,7 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-using Bodu.Test.Contracts;
-using Bodu.Test.Kat;
+using Bodu.IO.Hashing.Contracts;
 
 namespace Bodu.IO.Hashing.Checksums.Contracts;
