+++ Bodu.IO.Hashing/test/IO.Hashing/Fnv1a64ContractTests.cs
@@ -4,8 +4,6 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-using Bodu.Test.Contracts;
-using Bodu.Test.Kat;
 
 namespace Bodu.IO.Hashing.Contracts;
