+++ Bodu.IO.Hashing/test/IO.Hashing/NonCryptographicHashKnownAnswer.cs
@@ -4,6 +4,7 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
+
 using Bodu.Test.Kat;
 
 namespace Bodu.IO.Hashing;
